Add active/deactive project APIs #125

+++ Data.Abstractions/Models/IProject.cs
@@ -20,7 +20,7 @@ namespace Data.Abstractions.Models
 
         DateTime CreateDateTime { get; set; }
 
-        double Budget { get; set; }
+        int Budget { get; set; }
 
         double Weight { get; set; }
